Move Auth method to a concern
