+++ app/controllers/posts_controller.rb
@@ -1,4 +1,5 @@
 class PostsController < ApplicationController
+  include Secure
   before_action :authenticate_user!, only: [:create, :update]
 
   rescue_from Exception do |exception|
@@ -52,18 +53,5 @@ class PostsController < ApplicationController
 
   def update_params
     params.require(:post).permit(:title, :content, :published)
-  end
-  
-  def authenticate_user!    
-    token_regex = /Bearer (\w+)/    
-    headers = request.headers      
-    if headers['Authorization'].present? && headers['Authorization'].match(token_regex)
-      token = headers['Authorization'].match(token_regex)[1]      
-      if(Current.user = User.find_by_auth_token(token))
-        return
-      end
-    end
-
-    render json: {error: 'Unauthorized'}, status: :unauthorized
-  end
+  end    
 end
